+++ NEWS.md
@@ -1,3 +1,11 @@
+# linemap 0.3.0
+
+## Major change
+
+- Use terra raster as input
+- remove former example dataset
+
+
 # linemap 0.2.0
 
 ## Minor changes